requirements.txt
flask-restful==0.3.10
flask_sqlalchemy==3.0.5
flask_migrate==4.0.4
python-dotenv==1.0.0
psycopg2-binary==2.9.6
flask-marshmallow==0.15.0
flask-jwt-extended==4.5.2
marshmallow-sqlalchemy==0.29.0